copr: (refactor) Lift heap struct out from top_n_executor (#14096)
ref tikv/tikv#13936
